Module ktest::tree

The source module tree of ktests.

In the KtestTree, the root is abstract, and the children of the root are the crates. The leaves are the test functions. Nodes other than the root and the leaves are modules.

Structs

KtestCrate

KtestCrateIter

The KtestCrateIter will iterate over all modules in a crate. Yeilding KtestModules. The iterator will return modules in the depth-first-search order of the module tree.

KtestModule

KtestModuleIter

The KtestModuleIter will iterate over all tests in a crate. Yeilding KtestItems.

KtestTree

KtestTreeIter

The KtestTreeIter will iterate over all crates. Yeilding KtestCrates.